refactor(code-similarity): restructure test pyramid and harden edges
Move contract filtering and test threshold integration tests from
check.rs to tests/code_similarity/evaluation.rs. Remove duplicate
tests covered at the unit level. Fix is_cfg_test_attr false rejection
of compound cfg attributes. Add JS/TS extensions to CI and hook
checks. Bundle threshold params into ThresholdSet. Fix passive test
names, imprecise test name, missing Token docs, and broken handbook
link.
