+++ crates/scute-core/tests/code_similarity.rs
@@ -2,6 +2,8 @@
 mod clone_detection;
 #[path = "code_similarity/end_to_end.rs"]
 mod end_to_end;
+#[path = "code_similarity/evaluation.rs"]
+mod evaluation;
 #[path = "code_similarity/helpers.rs"]
 mod helpers;
 #[path = "code_similarity/normalization.rs"]
